+++ tests/integration_tests.rs
@@ -8,16 +8,34 @@ use rand::{rngs::SmallRng, Rng, SeedableRng};
 use compile_time_sort::{
     into_sorted_bool_array, into_sorted_char_array, into_sorted_f32_array, into_sorted_f64_array,
     into_sorted_i128_array, into_sorted_i16_array, into_sorted_i32_array, into_sorted_i64_array,
-    into_sorted_i8_array, into_sorted_isize_array, into_sorted_u128_array, into_sorted_u16_array,
-    into_sorted_u32_array, into_sorted_u64_array, into_sorted_u8_array, into_sorted_usize_array,
-    sort_f32_slice, sort_f64_slice,
+    into_sorted_i8_array, into_sorted_isize_array, into_sorted_str_array, into_sorted_u128_array,
+    into_sorted_u16_array, into_sorted_u32_array, into_sorted_u64_array, into_sorted_u8_array,
+    into_sorted_u8_slice_array, into_sorted_usize_array,
+};
+
+#[cfg(feature = "nested")]
+use compile_time_sort::{
+    into_sorted_bool_slice_array, into_sorted_i128_slice_array, into_sorted_i16_slice_array,
+    into_sorted_i32_slice_array, into_sorted_i64_slice_array, into_sorted_i8_slice_array,
+    into_sorted_isize_slice_array, into_sorted_u128_slice_array, into_sorted_u16_slice_array,
+    into_sorted_u32_slice_array, into_sorted_u64_slice_array, into_sorted_usize_slice_array,
+    sort_bool_slice_slice,
 };
 
 #[rustversion::since(1.83.0)]
 use compile_time_sort::{
-    sort_bool_slice, sort_char_slice, sort_i128_slice, sort_i16_slice, sort_i32_slice,
-    sort_i64_slice, sort_i8_slice, sort_isize_slice, sort_u128_slice, sort_u16_slice,
-    sort_u32_slice, sort_u64_slice, sort_u8_slice, sort_usize_slice,
+    sort_bool_slice, sort_char_slice, sort_f32_slice, sort_f64_slice, sort_i128_slice,
+    sort_i16_slice, sort_i32_slice, sort_i64_slice, sort_i8_slice, sort_isize_slice,
+    sort_str_slice, sort_u128_slice, sort_u16_slice, sort_u32_slice, sort_u64_slice, sort_u8_slice,
+    sort_u8_slice_slice, sort_usize_slice,
+};
+
+#[cfg(feature = "nested")]
+#[rustversion::since(1.83.0)]
+use compile_time_sort::{
+    sort_i128_slice_slice, sort_i16_slice_slice, sort_i32_slice_slice, sort_i64_slice_slice,
+    sort_i8_slice_slice, sort_isize_slice_slice, sort_u128_slice_slice, sort_u16_slice_slice,
+    sort_u32_slice_slice, sort_u64_slice_slice, sort_usize_slice_slice,
 };
 
 use paste::paste;
@@ -45,6 +63,11 @@ macro_rules! test_unsigned_integer {
             paste! {
                 #[test]
                 fn [<test_sort_ $tpe>]() {
+                    const TINY_ARR: [$tpe; 1] = [1];
+                    const SORTED_TINY_ARR: [$tpe; 1] = [<into_sorted_ $tpe _array>](TINY_ARR);
+                    assert!(SORTED_TINY_ARR.is_sorted());
+
+
                     const REV_ARRAY: [$tpe; 3] = [3, 2, 1];
                     const SORTED_REV_ARRAY: [$tpe; 3] = [<into_sorted_ $tpe _array>](REV_ARRAY);
                     assert!(SORTED_REV_ARRAY.is_sorted());
@@ -109,7 +132,19 @@ macro_rules! test_unsigned_integer {
                         arr
                     };
 
+                    const REVERSE_SORTED: [$tpe; 255] = {
+                        let mut arr = [0; 255];
+                        let mut i = arr.len();
+                        while i > 0 {
+                            arr[i-1] = i as $tpe;
+                            i -= 1
+                        }
+                        [<sort_ $tpe _slice>](&mut arr);
+                        arr
+                    };
+
                     assert!(BIG_IDENTICAL.is_sorted());
+                    assert!(REVERSE_SORTED.is_sorted());
                 }
             }
         )+
@@ -157,22 +192,17 @@ macro_rules! test_signed_integer {
                         quickcheck_case_2
                     };
 
-                    assert!(SORTED_QUICKCHECK_CASE_1.is_sorted());
-                    assert!(SORTED_QUICKCHECK_CASE_2.is_sorted());
-                }
+                    const SORTED_QUICKCHECK_CASE_3: [$tpe; 27] = {
+                        let mut quickcheck_case_3 = [0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0];
 
-                #[test]
-                #[rustversion::since(1.83.0)]
-                fn [<test_sort_ $tpe _slice>]() {
-                    const SORTED_QUICKCHECK_CASE_1: [$tpe; 27] = {
-                        let mut quickcheck_case_1 = [0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0];
+                        [<sort_ $tpe _slice>](&mut quickcheck_case_3);
 
-                        [<sort_ $tpe _slice>](&mut quickcheck_case_1);
-
-                        quickcheck_case_1
+                        quickcheck_case_3
                     };
 
                     assert!(SORTED_QUICKCHECK_CASE_1.is_sorted());
+                    assert!(SORTED_QUICKCHECK_CASE_2.is_sorted());
+                    assert!(SORTED_QUICKCHECK_CASE_3.is_sorted());
                 }
 
                 // Also run all the tests for unsigned integers on the signed integers
@@ -196,6 +226,139 @@ quickcheck_slice_sort! {
     usize, isize,
     char,
     bool
+}
+
+#[cfg(feature = "nested")]
+#[test]
+fn test_sort_bool_slice_arrays() {
+    const ARR: [&[bool]; 3] = [&[false, false], &[true], &[true, true]];
+    const SORTED_ARR: [&[bool]; 3] = into_sorted_bool_slice_array(ARR);
+
+    assert!(SORTED_ARR.is_sorted());
+}
+
+#[cfg(feature = "nested")]
+#[rustversion::since(1.83.0)]
+#[test]
+fn test_sort_bool_slice_slices() {
+    const SORTED_ARR: [&[bool]; 3] = {
+        let mut arr: [&[bool]; 3] = [&[false], &[true], &[true, true]];
+        sort_bool_slice_slice(&mut arr);
+        arr
+    };
+
+    assert!(SORTED_ARR.is_sorted());
+}
+
+macro_rules! test_unsigned_slices {
+    ($($tpe:ty),+) => {
+        $(
+            paste! {
+                #[test]
+                fn [<test_sort_ $tpe _slice_array>]() {
+                    const ARR: [&[$tpe]; 4] = [&[0, 1], &[0, 0], &[1, 0], &[1, 1]];
+                    const SORTED_ARR: [&[$tpe]; 4] = [<into_sorted_ $tpe _slice_array>](ARR);
+
+                    assert!(SORTED_ARR.is_sorted());
+                }
+
+                #[test]
+                fn [<test_sort_long_identical_ $tpe _slice_array>]() {
+                    const ARR: [&[$tpe]; 500] = [&[0]; 500];
+                    const SORTED_ARR: [&[$tpe]; 500] = [<into_sorted_ $tpe _slice_array>](ARR);
+
+                    assert!(SORTED_ARR.is_sorted());
+                }
+
+                #[rustversion::since(1.83.0)]
+                #[test]
+                fn [<test_sort_ $tpe _slice_slice>]() {
+                     const SORTED_ARR: [&[$tpe]; 4] = {
+                        let mut arr: [&[$tpe]; 4] = [&[0, 1], &[0, 0], &[1, 0], &[1, 1]];
+                        [<sort_ $tpe _slice_slice>](&mut arr);
+                        arr
+                    };
+
+                    assert!(SORTED_ARR.is_sorted());
+                }
+
+                #[rustversion::since(1.83.0)]
+                #[test]
+                fn [<test_sort_long_identical_ $tpe _slice_slice>]() {
+                    const SORTED_ARR: [&[$tpe]; 500] = {
+                        let mut arr: [&[$tpe]; 500] = [&[]; 500];
+                        [<sort_ $tpe _slice_slice>](&mut arr);
+                        arr
+                    };
+
+                    assert!(SORTED_ARR.is_sorted());
+                }
+            }
+        )+
+    };
+}
+
+#[cfg(feature = "nested")]
+macro_rules! test_signed_slices {
+    ($($tpe:ty),+) => {
+        $(
+            paste! {
+                #[test]
+                fn [<test_sort_ $tpe _slice_array_with_negatives>]() {
+                    const ARR: [&[$tpe]; 8] = [&[0, 1], &[0, 0], &[1, 0], &[1, 1], &[0, -1], &[0, 0], &[-1, 0], &[-1, -1]];
+                    const SORTED_ARR: [&[$tpe]; 8] = [<into_sorted_ $tpe _slice_array>](ARR);
+
+                    assert!(SORTED_ARR.is_sorted());
+                }
+
+                #[rustversion::since(1.83.0)]
+                #[test]
+                fn [<test_sort_ $tpe _slice_slice_with_negatives>]() {
+                     const SORTED_ARR: [&[$tpe]; 8] = {
+                        let mut arr: [&[$tpe]; 8] = [&[0, 1], &[0, 0], &[1, 0], &[1, 1], &[0, -1], &[0, 0], &[-1, 0], &[-1, -1]];
+                        [<sort_ $tpe _slice_slice>](&mut arr);
+                        arr
+                    };
+
+                    assert!(SORTED_ARR.is_sorted());
+                }
+            }
+        )+
+
+        test_unsigned_slices! { $($tpe),+ }
+    };
+}
+
+test_unsigned_slices! { u8 }
+
+#[cfg(feature = "nested")]
+test_unsigned_slices! {
+    u16, u32, u64, u128, usize
+}
+
+#[cfg(feature = "nested")]
+test_signed_slices! {
+    i8, i16, i32, i64, i128, isize
+}
+
+#[test]
+fn test_sort_str_array() {
+    const ARR: [&str; 4] = ["abc", "abd", "aaaaa", "l"];
+    const SORTED_ARR: [&str; 4] = into_sorted_str_array(ARR);
+
+    assert!(SORTED_ARR.is_sorted());
+}
+
+#[rustversion::since(1.83.0)]
+#[test]
+fn test_sort_str_slice() {
+    const SORTED_ARR: [&str; 4] = {
+        let mut arr: [&str; 4] = ["abc", "abd", "aaaaa", "l"];
+        sort_str_slice(&mut arr);
+        arr
+    };
+
+    assert!(SORTED_ARR.is_sorted());
 }
 
 #[test]
@@ -272,6 +435,10 @@ fn test_f32_sort_slice() {
     let mut random_array: [f32; 500] = core::array::from_fn(|_| rng.gen());
     sort_f32_slice(&mut random_array);
     assert!(random_array.is_sorted());
+
+    let mut all_same = vec![42.0; 500];
+    sort_f32_slice(&mut all_same);
+    assert!(all_same.is_sorted());
 }
 
 #[test]
@@ -333,8 +500,13 @@ fn test_f64_sort_slice() {
     let mut random_array: [f64; 500] = core::array::from_fn(|_| rng.gen());
     sort_f64_slice(&mut random_array);
     assert!(random_array.is_sorted());
+
+    let mut all_same = vec![42.0; 500];
+    sort_f64_slice(&mut all_same);
+    assert!(all_same.is_sorted());
 }
 
+#[rustversion::since(1.83.0)]
 quickcheck! {
     fn quickcheck_f32_slice(vec: Vec<f32>) -> bool {
         let mut vec = vec;